chore: enhance testing CI/CD  (#117)
- copied over the pytest config from med-imagetools
- can easily view errors in gha during pytest [without going into the
action
logs](https://github.com/bhklab/med-imagetools/actions/runs/13116280733/attempts/1#summary-36591278874)

- small fix to the fixtures that delete temporary paths and are used by
multiple tests (can cause errors during parallel runs)

- also cleanups for each PRs gha caches

+++ tests/test_feature_extraction.py
@@ -44,23 +44,27 @@ def lung4DRTSTRUCTImage():
 def pyradiomicsParamFilePath():
     return "src/readii/data/default_pyradiomics.yaml"
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def nsclcMetadataPath():
     oldpath = Path("tests/output/ct_to_seg_match_list_NSCLC_Radiogenomics.csv")
     newpath = Path("tests/NSCLC_Radiogenomics/procdata/ct_to_seg_match_list_NSCLC_Radiogenomics.csv")
     newpath.parent.mkdir(parents=True, exist_ok=True)
     shutil.copy(oldpath, newpath)
     yield newpath.as_posix()
-    newpath.unlink()
 
-@pytest.fixture
+    if newpath.exists():
+        newpath.unlink()
+
+@pytest.fixture(scope="module")
 def lung4DMetadataPath():
     oldpath = Path("tests/output/ct_to_seg_match_list_4D-Lung.csv")
     newpath = Path("tests/4D-Lung/procdata/ct_to_seg_match_list_4D-Lung.csv")
     newpath.parent.mkdir(parents=True, exist_ok=True)
     shutil.copy(oldpath, newpath)
     yield newpath.as_posix()
-    newpath.unlink()
+
+    if newpath.exists():
+        newpath.unlink()
 
 
 def test_singleRadiomicFeatureExtraction_SEG(nsclcCTImage, nsclcSEGImage, pyradiomicsParamFilePath):
